department is ok in backend

+++ routes/voteItems.js
@@ -2,15 +2,22 @@ var express = require('express');
 var router = express.Router();
 var voteItem_sc = require("../Schema/voteItems");
 var auth = require('../tools/authentication');
+var {checkPermissions}=require("../tools/auth");
+
 
 
 //Add new program or channel
 router.post('/new', auth,function (req, res) {
-  var allowedPermissions=[122]
+  if(req.session.type!=0){
+    var allowedPermissions=[122]
+    if(!checkPermissions(allowedPermissions,req.session.permissions))return res.status(403).json({error:"You don't have access to this api."})
+    }
   
-  // console.log('Now U can save a new program or channel (voteItems)...');
-  var voteItem = new voteItem_sc(req.body);
-  console.log(req.body);
+  var {title,text,keyboard}=req.body;
+
+  var voteItem = new voteItem_sc({title,type,description});
+  voteItem.departmentId=req.session.departmentId;
+
   voteItem.save(function (err, result) {
 
     if (!err) {
@@ -29,9 +36,16 @@ router.post('/new', auth,function (req, res) {
 
 //Get all programs and channels (voteItems) which are enable (not deleted)
 router.post('/all', auth,function (req, res) {
-  var allowedPermissions=[122]
+  if(req.session.type!=0){
+    var allowedPermissions=[122]
+    if(!checkPermissions(allowedPermissions,req.session.permissions))return res.status(403).json({error:"You don't have access to this api."})
+    }
   
-  voteItem_sc.find({enable:1}, function (err, result) {
+  var data;
+  if (req.session.type < 2) data = {};
+  else data = {$and:[{'departmentId': req.session.departmentId},{enable:1}]};
+  
+  voteItem_sc.find(data, function (err, result) {
     if (!err) {
       if (result) {
         res.json({
@@ -52,8 +66,8 @@ router.post('/all', auth,function (req, res) {
 
 //Get all programs and channels (voteItems) ENABLE+DISABLE for recovery
 router.post('/all/recover', auth,function (req, res) {
-  var allowedPermissions=[122]
   
+  if(req.session.type!=0) return res.status(403).json({error:"You don't have access to this api."})
   voteItem_sc.find({}, function (err, result) {
     if (!err) {
       if (result) {
@@ -98,7 +112,11 @@ router.post('/all/recover', auth,function (req, res) {
 
 //update voteItems (by id)
 router.post('/update',auth,function(req,res){
-  var allowedPermissions=[122]
+  if(req.session.type!=0){
+    var allowedPermissions=[122]
+    if(!checkPermissions(allowedPermissions,req.session.permissions))return res.status(403).json({error:"You don't have access to this api."})
+    }
+  
   
   console.log('query:',req.body)
   voteItem_sc.findById(req.body._id).exec(function(err,result){
@@ -109,6 +127,7 @@ router.post('/update',auth,function(req,res){
       result.description=req.body.description||result.description;
       result.channelId=req.body.channelId||result.channelId;
       result.personnels=req.body.personnels||result.personnels;
+      result.departmentId=req.body.departmentId||result.departmentId;
 
       // Save the updated document back to the database
       result.save(function(err,result){
@@ -132,7 +151,10 @@ router.post('/update',auth,function(req,res){
 //INPUT:{"_id":"5a1e711ed411741d84d10a29"}
 
 router.post('/disable',auth, function (req, res) {
+  if(req.session.type!=0){  
   var allowedPermissions=[122]
+  if(!checkPermissions(allowedPermissions,req.session.permissions))return res.status(403).json({error:"You don't have access to this api."})
+  }
   
   console.log('query', req.body)
     voteItem_sc.findById(req.body._id).exec(function (err, result) {
@@ -163,10 +185,12 @@ router.post('/disable',auth, function (req, res) {
 //INPUT:{"_id":"5a1e711ed411741d84d10a29"}
 
 router.post('/recover', auth,function (req, res) {
-  console.log('query', req.body)
+  if(req.session.type!=0)  
+    return res.status(403).json({error:"You don't have access to this api."})
+    
     voteItem_sc.findById(req.body._id).exec(function (err, result) {
     if (!err) {
-      console.log("voteItem:",result) 
+       
       result.enable=1;  
       result.save(function(err,result){
         if(!err){
@@ -193,7 +217,8 @@ router.post('/recover', auth,function (req, res) {
 //INPUT:{"_id":"5a1e711ed411741d84d10a29"}
 
 router.post('/delete',auth, function (req, res) {
-  console.log('query', req.body)
+  if(req.session.type!=0)  return res.status(403).json({error:"You don't have access to this api."})
+  
     voteItem_sc.findByIdAndRemove(req.body._id).exec(function (err, result) {
     if (!err) {
       res.status(200);
@@ -205,8 +230,5 @@ router.post('/delete',auth, function (req, res) {
     }
   })
 })
-
-
-
 
 module.exports = router;
